New changes!
I realized our method of passing an identity matrix through the functions to get out the transformation matrix was not working as we wanted. For example, the matrix actually used in a translation is of a 4x4 format, and the one we got out was a 3x3. Now each of the scale, translation, and rotation functions return a transformed object and the matrix that was used to do the transformation. I standardized our variable names, function names, comment style, and whitespace style.